Inline util into server.tsx and add unit test
